pyproject.toml
[tool.poetry]
name = "rss-reeder"
version = "0.1.0"
description = ""
authors = ["dmitry.tyutryumov"]
license = "MIT"
readme = "README.md"
packages = [{include = "rss_reeder"}]

[tool.poetry.dependencies]
python = "^3.10"
feedparser = "^6.0.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.24.2"
psycopg2-binary = "^2.9.5"
celery = "^5.2.7"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pytest = "^7.2.0"
pre-commit = "^2.20.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-xdist = "^3.0.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
multi_line_output = 3
force_grid_wrap = 2
include_trailing_comma = "True"
ensure_newline_before_comments = "True"
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | docs
  | docker
)/
'''
[tool.isort]
profile = "black"
